fix: CheckoutDialog を分割してメンテナビリティ向上
- checkout-dialog.tsx: 外枠・状態管理のみ（約240行）
- payment-cash-tab.tsx: 現金支払UI
- payment-card-tab.tsx: カード/QR支払UI（mode propで切替）
- use-checkout.ts: ビジネスロジックカスタムフック
- 既存テスト全14件が引き続きパス

Closes #482
